+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [7.1.3] - 2021-11-12
 ### Added
  - Run test suite with PHP 8.1
 
@@ -541,7 +543,8 @@ See :
  * https://github.com/janephp/jane/releases
  * https://github.com/janephp/openapi/releases
 
-[Unreleased]: https://github.com/janephp/janephp/compare/v7.1.2...HEAD
+[Unreleased]: https://github.com/janephp/janephp/compare/v7.1.3...HEAD
+[7.1.3]: https://github.com/janephp/janephp/compare/v7.1.2...v7.1.3
 [7.1.2]: https://github.com/janephp/janephp/compare/v7.1.1...v7.1.2
 [7.1.1]: https://github.com/janephp/janephp/compare/v7.1.0...v7.1.1
 [7.1.0]: https://github.com/janephp/janephp/compare/v7.0.0...v7.1.0